Testeados los endpoints básicos. Modificados para que devuelvan la info esperada.

+++ API/WCA.Domain/Entities/CafeLote.cs
@@ -14,7 +14,7 @@
         public int VariedadId { get; set; }
         public int TuesteId { get; set; }
 
-        public double AltitudMedia { get; set; } //columna computada
+        public decimal AltitudMedia { get; set; } //columna computada
 
         public string? DescripcionExtendida { get; set; }
